==> 2-Processing/1.DEGAnalysis.Rmd <==
---
title: "1.DGEAnalysis"
author: "Athmaja Viswanath"
date: "2024-09-19"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
```

#0. LOADING REQUIRED PACKAGES
```{r loading packages}
library(DESeq2)
library(ggplot2)
library(cowplot) #add on to ggplot for better themes and figure customization
library(lemon) #to work with legends and axes in ggplot2
library(dplyr)
library(gdata)
library(RColorBrewer)
library(colorBlindness)
library(colorspace)
library(tidyverse)

theme_set(theme_classic())
```

#1. DIFFERENTIAL GENE EXPRESSION FOR ORTHOLOGOUS GENES
##A Data preparation 
###I. Separating orthologous gene names so that they can be used to filter readcounts for each species.
Total number of orthologous genes detected is 13785 genes across C. remanei and C. latens (gene list from Daniel)
```{r}
# Load ortholog data from a tab-separated file, ignoring comments
orthologs = read.table("../1-Input/new_1to1_orthologgenelist.txt", sep="\t", header=TRUE, comment.char="#")

# Extract the relevant columns into separate variables
cre_ortho = orthologs[, 1]            # First column: cre_ortho, Cre names
clat_ortho = orthologs[, 2]           # Second column: clat_ortho, Clat names
cumulative_ortho = orthologs[, 3]     # Third column: cumulative_ortho, joint names

# Display the first few rows of the data frames
head(orthologs)
head(cre_ortho)
head(clat_ortho)

##Number of genes 
nrow(orthologs)
length(cre_ortho)
length(clat_ortho)
```

###II Filtering readcoutns for all samples to retain orthologous genes and combining to make one readcount file
Total number of genes in C. remanei = 20091
Total number of gene sin C. latens = 23073
Total number of 1:1 orthologous genes = 13785

```{r}
#########
#C.remanei
#########
# Load counts for C. remanei
cre_counts = read.table("../1-Input/C.remanei_cumulative.txt", sep="\t", header=TRUE, row.names=1, comment.char="#")
# Look at the structure of data
head(cre_counts)
colnames(cre_counts)
nrow(cre_counts) #20091
# Filter to include only orthologs and exclude Whole animal samples
cre_counts_ortho = cre_counts[cre_ortho, -c(13:15)] 
# Update row names to match cumulative orthologs
rownames(cre_counts_ortho) = cumulative_ortho 
# Check the number of rows and column names
nrow(cre_counts_ortho)
colnames(cre_counts_ortho)

#########
#C.latens
#########
# Load counts for C. latens
clat_counts = read.table("../1-Input/C.latens_cumulative.txt", sep="\t", header=TRUE, row.names=1, comment.char="#")
# Look at the structure of data
head(clat_counts)
colnames(clat_counts)
nrow(clat_counts) #23073
# Filter to include only orthologs and exclude Whole animal samples
clat_counts_ortho = clat_counts[clat_ortho, -c(13:15)]
# Update row names to match cumulative orthologs
rownames(clat_counts_ortho) = cumulative_ortho
# Check the number of rows and column names for C. latens
nrow(clat_counts_ortho)
names(clat_counts_ortho)


# Combine ortholog counts from C. remanei and C. latens
wt_counts_orthologs = cbind(cre_counts_ortho, clat_counts_ortho)

# Display current column names
colnames(wt_counts_orthologs)

# Clean up column names by removing the "X" prefix
colnames(wt_counts_orthologs) = gsub("X", "", colnames(wt_counts_orthologs))
colnames(wt_counts_orthologs)
```

##B Performing differnetial gene expression analysis
###I Preparing coldata
```{r}
# Prepare the coldata data frame
sample_name = colnames(wt_counts_orthologs)  # Get sample names
tissue = substr(sample_name, 3, 3)            # Extract tissue type from sample names
sex = substr(sample_name, 2, 2)               # Extract sex from sample names
species = c(rep("Cre", 12), rep("Clat", 12))  # Assign species labels
batch = rep(c(1, 2, 3),8)                # Assign batch numbers

# Create the coldata data frame
coldata = data.frame(sample_name, species, sex, tissue, batch)

# Convert columns to factors for analysis
coldata$species = factor(coldata$species)
coldata$tissue = factor(coldata$tissue)
coldata$sex = factor(coldata$sex)
coldata$batch = factor(coldata$batch)

# Display the coldata data frame
coldata  # Note: WM samples are not included
```

###II Performing differential gene expression using DESeq2 and extracting model matrix
```{r}
#Run deseq2 analysis 
dds_wt = DESeqDataSetFromMatrix(
  countData = wt_counts_orthologs, 
  colData = coldata, 
  design = ~ species + tissue + sex + species:tissue + sex:tissue + species:sex)

# Perform the DESeq analysis
dds_wt = DESeq(dds_wt)

# Display the names of the results generated by the model
resultsNames(dds_wt)
summary(results(dds_wt))

# Create the model matrix based on the design
mod_mat_wt <- model.matrix(design(dds_wt), colData(dds_wt))
```

###III Extracting constrasts/comparisons from the main model
####i BETWEEN SPECIES
Identifying species-biased genes (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM)

Results: From 13,503 gene orthologs with non-zero read counts, we found ~49% of genes (nDEG = 6563) to show differential expression between species, being evenly divided between C. remanei and C. latens having higher expression in one species or the other (nCrem = 3224, nClat = 3339), with the remaining 6865 genes showing conserved expression between the species (Figure 3.1A). 
```{r}
#Define coefficient vectors for each condition
cre = colMeans(mod_mat_wt[dds_wt$species == "Cre", ])
clat = colMeans(mod_mat_wt[dds_wt$species == "Clat", ])

# Perform contrast analysis between C. remanei and C. latens
crexclat = results(dds_wt, contrast = cre - clat, alpha = 0.05)  # Setting C. latens as the baseline

# Combine results with a new column indicating significance and direction of change
crexclat_res <- cbind(
  as.data.frame(crexclat), 
  as.data.frame(crexclat) %>% 
    mutate(crexclat = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(crexclat)
  )
# Remove rows with NA values
crexclat_res = na.omit(crexclat_res)

# Separate results into up/down regulated and conserved
crexclat_res1 = crexclat_res %>% filter(crexclat_res$crexclat != 0) # Upregulated and downregulated genes
crexclat_res2 = crexclat_res %>% filter(crexclat_res$crexclat == 0) #conserved expression genes

# Summarize the results
summary(crexclat)# Summary of results; upregulated in C. remanei with C. latens as the reference
nrow(crexclat_res) # Total Number of genes tested
table(crexclat_res$crexclat) # Count of species-biased and conserved genes

#Saving results
#write.csv(crexclat_res, file = "../3-Output/crexclat_genes.csv")
```

####ii BETWEEN TISSUES
FINDING GENES THAT ARE TISSUE-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM)

```{r}

#Define coefficient vectors for each condition
gonad = colMeans(mod_mat_wt[dds_wt$tissue == "G", ])
soma = colMeans(mod_mat_wt[dds_wt$tissue == "S", ])

# Perform contrast analysis comparing soma to gonad
gxs = results(dds_wt, contrast = soma - gonad, alpha = 0.05)  # Setting gonad as the baseline

# Combine results with a new column indicating significance and direction of change
gxs_res <- cbind(
  as.data.frame(gxs), 
  as.data.frame(gxs) %>% 
    mutate(gxs = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(gxs)
  )

# Remove rows with NA values
gxs_res = na.omit(gxs_res)

# Separate results into significant (up/downregualted) and non-significant (tissue neutral)
gxs_res1 = gxs_res %>% filter(gxs_res$gxs != 0) #tissue-biased genes
gxs_res2 = gxs_res %>% filter(gxs_res$gxs == 0) #tissue-neutral genes

# Summary of the results
nrow(gxs_res) #number of genes tested
summary(gxs)# Summary of results; upregulated in soma with gonad as the reference
table(gxs_res$gxs)

# Display the design formula used in DESeq2
design(dds_wt)

# Optionally, save results to a CSV file
#write.csv(gxs_res, file = "../3-Output/gonadxsoma_genes.csv")
```

####iii BETWEEN SEXES
FINDING GENES THAT ARE SEX-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM) 

```{r}
#Define coefficient vectors for each sex condition
male = colMeans(mod_mat_wt[dds_wt$sex == "M", ])
female = colMeans(mod_mat_wt[dds_wt$sex == "F", ])

# Perform contrast analysis comparing males to females
mxf = results(dds_wt, contrast = male - female, alpha = 0.05) # Setting female as the baseline

# Combine results with a new column indicating significance and direction of change
mxf_res <- cbind(
  as.data.frame(mxf), 
  as.data.frame(mxf) %>% 
    mutate(mxf = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(mxf)
  )

# Remove rows with NA values
mxf_res = na.omit(mxf_res)

# Separate results into sex-biased and sex-neutral genes
mxf_res1 = mxf_res %>% filter(mxf_res$mxf != 0)
mxf_res2 = mxf_res %>% filter(mxf_res$mxf == 0)

# Summary of the results
nrow(mxf_res) #number of genes tested
summary(mxf)# Summary of results; upregulated in males with females as the reference
table(mxf_res$mxf)

# Save results to a CSV file
#write.csv(mxf_res, file = "../3-Output/mxf_genes.csv")
```

```{r}
#####COMBINING DATA
# Combine results from different analyses into a single data frame
all_total_biasgene_info = cbind(crexclat_res, gxs_res, mxf_res)  # Includes DEGs with interaction information

# Create a contingency table to summarize the relationships between the datasets
table(all_total_biasgene_info$crexclat, all_total_biasgene_info$gxs)  # First dataset as row names

table(all_total_biasgene_info$crexclat, all_total_biasgene_info$mxf) # First dataset as row names

```

####iv EFFECTS OF SPECIES & TISSUE DIFFERENCES
```{r}
# Perform contrast analysis 
spxts = results(dds_wt, contrast = list("speciesCre.tissueS"), alpha = 0.05) # Baseline is C. latens

# Combine results and create a new column indicating significance and direction of change
spxts_res <- cbind(
  as.data.frame(spxts), 
  as.data.frame(spxts) %>% 
    mutate(spxts = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(spxts)
  )

# Remove rows with NA values
spxts_res = na.omit(spxts_res)

# Separate significant and non-significant results
spxts_res1 = spxts_res %>% filter(spxts_res$spxts != 0 ) # Significant results
spxts_res2 = spxts_res %>% filter(spxts_res$spxts == 0 ) # 

# Summarize the results
summary(spxts)  # Summary of the analysis results
nrow(spxts_res)  # Number of genes
table(spxts_res$spxts)
```

####v EFFECTS OF SPECIES & SEX DIFFERENCES

```{r}
#Perform contrast analysis for species and sex differences
spxsex = results(dds_wt, contrast = list("speciesCre.sexM"), alpha = 0.05) # Baseline is C. latens

# Combine results and add a column indicating significance and direction of change
spxsex_res <- cbind(
  as.data.frame(spxsex), 
  as.data.frame(spxsex) %>% 
    mutate(spxsex = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(spxsex)
  )

# Remove rows with NA values
spxsex_res = na.omit(spxsex_res)

# Separate significant and non-significant results
spxsex_res1 = spxsex_res %>% filter(spxsex_res$spxsex != 0 ) # Significant results
spxsex_res2 = spxsex_res %>% filter(spxsex_res$spxsex == 0 ) # Non-Significant results

# Summarize the results
summary(spxsex)  # Summary of the analysis results
nrow(spxsex_res)  # Number of genes
table(spxsex_res$spxsex)
```



####vi EFFECTS OF TISSUE & SEX DIFFERENCES

```{r}
# Perform contrast analysis 
tsxsex = results(dds_wt, contrast = list("tissueS.sexM"), alpha = 0.05) #Baseline C. latens

#Combine results and add a column indicating significance and direction of change
tsxsex_res <- cbind(
  as.data.frame(tsxsex), 
  as.data.frame(tsxsex) %>% 
    mutate(tsxsex = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(tsxsex)
  )

tsxsex_res = na.omit(tsxsex_res)
tsxsex_res1 = tsxsex_res %>% filter(tsxsex_res$tsxsex != 0 )
tsxsex_res2 = tsxsex_res %>% filter(tsxsex_res$tsxsex == 0 )


#Summary of results
summary(tsxsex)
nrow(tsxsex_res)
table(tsxsex_res$tsxsex)
```

#2. DIFFERENTIAL GENE EXPRESSION ANALYSIS FOR C.remanei SPECIFIC GENES
##A Data preparation 
###I. Loading required data
```{r}
# Load ortholog data from a tab-separated file, ignoring comments
cre_specific = read.table("../1-Input/C_remanei_specific_genes_new_annotations.txt", sep="\t", header=FALSE, comment.char="#")

# Display the first few rows of the data frames
head(cre_specific)

#Renaming entries
cre_specific$V1 <- gsub("-R.*$", "", cre_specific$V1)

##Number of genes 
nrow(cre_specific) #982 GENES
length(cre_specific$V1) #982 GENES

```

###II Filtering readcoutns for all samples to retain Cre specific genes and combining to make one readcount file
Total number of genes in C. remanei = 20091
Total number of gene sin C. latens = 23073 
Total number of 1:1 orthologous genes = 13785 
Total number of C. remanei specific genes= 982 

```{r}
# Load counts for C. remanei
cre_counts = read.table("../1-Input/C.remanei_cumulative.txt", sep="\t", header=TRUE, row.names=1, comment.char="#")
# Look at the structure of data
head(cre_counts)
colnames(cre_counts)
nrow(cre_counts) #20091

# Filter to include only Cre specific genes and exclude Whole animal samples
counts_cre = cre_counts[cre_specific$V1, -c(13:15)] 

# Check the number of rows and column names
nrow(counts_cre) #982 genes
colnames(counts_cre)

# Clean up column names by removing the "X" prefix
colnames(counts_cre) = gsub("X", "", colnames(counts_cre))
colnames(counts_cre)
```
##B Performing differEnetial gene expression analysis
###I Preparing coldata
```{r}
# Prepare the coldata data frame
sample_name = colnames(counts_cre)  # Get sample names
tissue = substr(sample_name, 3, 3)            # Extract tissue type from sample names
sex = substr(sample_name, 2, 2)               # Extract sex from sample names
species = c(rep("Cre", 12))  # Assign species labels
batch = rep(c(1, 2, 3),4)                # Assign batch numbers

# Create the coldata data frame
coldata_cre = data.frame(sample_name, sex, tissue, batch)

# Convert columns to factors for analysis
#coldata_cre$species = factor(coldata_cre$species)
coldata_cre$tissue = factor(coldata_cre$tissue)
coldata_cre$sex = factor(coldata_cre$sex)
coldata_cre$batch = factor(coldata_cre$batch)

# Display the coldata data frame
coldata_cre  # Note: WM samples are not included
```

###II Performing differential gene expression using DESeq2 and extracting model matrix
```{r}
#Run deseq2 analysis 
dds_cre = DESeqDataSetFromMatrix(
  countData = counts_cre, 
  colData = coldata_cre, 
  design = ~ tissue + sex + sex:tissue)

# Perform the DESeq analysis
dds_cre = DESeq(dds_cre)

# Display the names of the results generated by the model
resultsNames(dds_cre)
summary(results(dds_cre))
# Create the model matrix based on the design
mod_mat_cre <- model.matrix(design(dds_cre), colData(dds_cre))

```
###III Extracting constrasts/comparisons from the main model
####i BETWEEN TISSUES
FINDING GENES THAT ARE TISSUE-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM)
```{r}
#Define coefficient vectors for each condition
gonad_cre = colMeans(mod_mat_cre[dds_cre$tissue == "G", ])
soma_cre = colMeans(mod_mat_cre[dds_cre$tissue == "S", ])

# Perform contrast analysis comparing soma to gonad
gxs_cre = results(dds_cre, contrast = soma_cre - gonad_cre, alpha = 0.05)  # Setting gonad as the baseline

# Combine results with a new column indicating significance and direction of change
gxs_res_cre <- cbind(
  as.data.frame(gxs_cre), 
  as.data.frame(gxs_cre) %>% 
    mutate(gxs = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(gxs)
  )

# Remove rows with NA values
gxs_res_cre = na.omit(gxs_res_cre)

# Separate results into significant (up/downregualted) and non-significant (tissue neutral)
gxs_res1_cre = gxs_res_cre %>% filter(gxs_res_cre$gxs != 0) #tissue-biased genes
gxs_res2_cre = gxs_res_cre %>% filter(gxs_res_cre$gxs == 0) #tissue-neutral genes

# Summary of the results
nrow(gxs_res_cre) #number of genes tested
summary(gxs_cre)# Summary of results; upregulated in soma with gonad as the reference
table(gxs_res_cre$gxs)

#Converting table to dataframe
cre_gxs_table = as.data.frame(table(gxs_res_cre$gxs)) %>%
  rename(Bias = Var1, genes_counts = Freq) %>%
  mutate(Bias = gsub("-1", "Gonad-biased", Bias), 
         Bias = gsub("1", "Soma-biased", Bias), 
         Bias = gsub("0", "Tissue-Neutral", Bias)) %>%
  mutate(Species = c(rep("Cre_TB", 3))) %>%
  mutate(Proportion = genes_counts / sum(genes_counts))

# Display the design formula used in DESeq2
design(dds_cre)

# Optionally, save results to a CSV file
# write.csv(gxs_res_cre, file = "../3-Output/DEGAnalysis/gonadxsoma_genes_crespecificgenes.csv")

```

####ii BETWEEN SEXES
FINDING GENES THAT ARE SEX-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM) 
```{r}
#Define coefficient vectors for each sex condition
male_cre = colMeans(mod_mat_cre[dds_cre$sex == "M", ])
female_cre = colMeans(mod_mat_cre[dds_cre$sex == "F", ])

# Perform contrast analysis comparing males to females
mxf_cre = results(dds_cre, contrast = male_cre - female_cre, alpha = 0.05) # Setting female as the baseline

# Combine results with a new column indicating significance and direction of change
mxf_res_cre <- cbind(
  as.data.frame(mxf_cre), 
  as.data.frame(mxf_cre) %>% 
    mutate(mxf = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(mxf)
  )

# Remove rows with NA values
mxf_res_cre = na.omit(mxf_res_cre)

# Separate results into sex-biased and sex-neutral genes
mxf_res1_cre = mxf_res_cre %>% filter(mxf_res_cre$mxf != 0)
mxf_res2_cre = mxf_res_cre %>% filter(mxf_res_cre$mxf == 0)

# Summary of the results
nrow(mxf_res_cre) #number of genes tested
summary(mxf_cre)# Summary of results; upregulated in males with females as the reference
table(mxf_res_cre$mxf)

#Converting table to dataframe
cre_mxf_table = as.data.frame(table(mxf_res_cre$mxf)) %>%
  rename(Bias = Var1, genes_counts = Freq) %>%
  mutate(Bias = gsub("-1", "Female-biased", Bias), 
         Bias = gsub("1", "Male-biased", Bias), 
         Bias = gsub("0", "Sex-Neutral", Bias)) %>%
  mutate(Species = c(rep("Cre_SB", 3))) %>%
  mutate(Proportion = genes_counts / sum(genes_counts))

# Save results to a CSV file
# write.csv(mxf_res_cre, file = "../3-Output/DEGAnalysis/mxf_genes_crespecificgenes.csv")
```

####iii EFFECTS OF TISSUE & SEX DIFFERENCES
```{r}
# Perform contrast analysis 
tsxsex_cre = results(dds_cre, contrast = list("tissueS.sexM"), alpha = 0.05) #Baseline female gonad

#Combine results and add a column indicating significance and direction of change
tsxsex_res_cre <- cbind(
  as.data.frame(tsxsex_cre), 
  as.data.frame(tsxsex_cre) %>% 
    mutate(tsxsex_cre = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(tsxsex_cre)
  )

tsxsex_res_cre = na.omit(tsxsex_res_cre)
tsxsex_res1_cre = tsxsex_res_cre %>% filter(tsxsex_res_cre$tsxsex != 0 )
tsxsex_res2_cre = tsxsex_res_cre %>% filter(tsxsex_res_cre$tsxsex == 0 )


#Summary of results
summary(tsxsex_cre)
nrow(tsxsex_res_cre)
table(tsxsex_res_cre$tsxsex)
```

#3. DIFFERENTIAL GENE EXPRESSION ANALYSIS FOR C.latens SPECIFIC GENES
##A Data preparation 
###I. Loading requried data
```{r}
# Load ortholog data from a tab-separated file, ignoring comments
clat_specific = read.table("../1-Input/C_latens_specific_genes_new_annotations.txt", sep="\t", header=FALSE, comment.char="#")

# Display the first few rows of the data frames
head(clat_specific)

#Renaming entries
clat_specific$V1 <- gsub("-R.*$", "", clat_specific$V1)

##Number of genes 
nrow(clat_specific) #2279 GENES
length(clat_specific$V1) #2279 GENES

```

###II Filtering readcoutns for all samples to retain Cre specific genes and combining to make one readcount file
Total number of genes in C. remanei = 20091
Total number of gene sin C. latens = 23073
Total number of 1:1 orthologous genes = 13785
Total number of C. remanei specific genes= 982
Total number of C. latens specific genes= 2279

```{r}
# Load counts for C. remanei
clat_counts = read.table("../1-Input/C.latens_cumulative.txt", sep="\t", header=TRUE, row.names=1, comment.char="#")
# Look at the structure of data
head(clat_counts)
colnames(clat_counts)
nrow(clat_counts) #23073

# Filter to include only clat specific genes and exclude Whole animal samples
counts_clat = clat_counts[clat_specific$V1, -c(13:15)] 

# Check the number of rows and column names
nrow(counts_clat) #2279
colnames(counts_clat)

# Clean up column names by removing the "X" prefix
colnames(counts_clat) = gsub("X", "", colnames(counts_clat))
colnames(counts_clat)
```

##B Performing differenetial gene expression analysis
###I Preparing coldata
```{r}
# Prepare the coldata data frame
sample_name = colnames(counts_clat)  # Get sample names
tissue = substr(sample_name, 3, 3)            # Extract tissue type from sample names
sex = substr(sample_name, 2, 2)               # Extract sex from sample names
species = c(rep("clat", 12))  # Assign species labels
batch = rep(c(1, 2, 3),4)                # Assign batch numbers

# Create the coldata data frame
coldata_clat = data.frame(sample_name, sex, tissue, batch)

# Convert columns to factors for analysis
#coldata_clat$species = factor(coldata_clat$species)
coldata_clat$tissue = factor(coldata_clat$tissue)
coldata_clat$sex = factor(coldata_clat$sex)
coldata_clat$batch = factor(coldata_clat$batch)

# Display the coldata data frame
coldata_clat  # Note: WM samples are not included
```

###II Performing differential gene expression using DESeq2 and extracting model matrix
```{r}
#Run deseq2 analysis 
dds_clat = DESeqDataSetFromMatrix(
  countData = counts_clat, 
  colData = coldata_clat, 
  design = ~ tissue + sex + sex:tissue)

# Perform the DESeq analysis
dds_clat = DESeq(dds_clat)

# Display the names of the results generated by the model
resultsNames(dds_clat)
summary(results(dds_clat))

# Create the model matrix based on the design
mod_mat_clat <- model.matrix(design(dds_clat), colData(dds_clat))

```
###III Extracting constrasts/comparisons from the main model
####i BETWEEN TISSUES
FINDING GENES THAT ARE TISSUE-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM)
```{r}

#Define coefficient vectors for each condition
gonad_clat = colMeans(mod_mat_clat[dds_clat$tissue == "G", ])
soma_clat = colMeans(mod_mat_clat[dds_clat$tissue == "S", ])

# Perform contrast analysis comparing soma to gonad
gxs_clat = results(dds_clat, contrast = soma_clat - gonad_clat, alpha = 0.05)  # Setting gonad as the baseline

# Combine results with a new column indicating significance and direction of change
gxs_res_clat <- cbind(
  as.data.frame(gxs_clat), 
  as.data.frame(gxs_clat) %>% 
    mutate(gxs = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(gxs)
  )

# Remove rows with NA values
gxs_res_clat = na.omit(gxs_res_clat)

# Separate results into significant (up/downregualted) and non-significant (tissue neutral)
gxs_res1_clat = gxs_res_clat %>% filter(gxs_res_clat$gxs != 0) #tissue-biased genes
gxs_res2_clat = gxs_res_clat %>% filter(gxs_res_clat$gxs == 0) #tissue-neutral genes

# Summary of the results
nrow(gxs_res_clat) #number of genes tested
summary(gxs_clat)# Summary of results; upregulated in soma with gonad as the reference
table(gxs_res_clat$gxs)

#Creating a dataframe form table
clat_gxs_table = as.data.frame(table(gxs_res_clat$gxs)) %>%
  rename(Bias = Var1, genes_counts = Freq) %>%
  mutate(Bias = gsub("-1", "Gonad-biased", Bias), 
         Bias = gsub("1", "Soma-biased", Bias), 
         Bias = gsub("0", "Tissue-Neutral", Bias)) %>%
  mutate(Species = c(rep("Clat_TB", 3))) %>%
  mutate(Proportion = genes_counts / sum(genes_counts))

# Display the design formula used in DESeq2
design(dds_clat)

# Optionally, save results to a CSV file
# write.csv(gxs_res_clat, file = "../3-Output/DEGAnalysis/gonadxsoma_genes_clatspecificgenes.csv")
```

####ii BETWEEN SEXES
FINDING GENES THAT ARE SEX-BIASED (INCLUDING THOSE THAT ARE DUE TO INTERATION TERM) 

```{r}
#Define coefficient vectors for each sex condition
male_clat = colMeans(mod_mat_clat[dds_clat$sex == "M", ])
female_clat = colMeans(mod_mat_clat[dds_clat$sex == "F", ])

# Perform contrast analysis comparing males to females
mxf_clat = results(dds_clat, contrast = male_clat - female_clat, alpha = 0.05) # Setting female as the baseline

# Combine results with a new column indicating significance and direction of change
mxf_res_clat <- cbind(
  as.data.frame(mxf_clat), 
  as.data.frame(mxf_clat) %>% 
    mutate(mxf = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(mxf)
  )

# Remove rows with NA values
mxf_res_clat= na.omit(mxf_res_clat)

# Separate results into sex-biased and sex-neutral genes
mxf_res1_clat = mxf_res_clat %>% filter(mxf_res_clat$mxf != 0)
mxf_res2_clat = mxf_res_clat %>% filter(mxf_res_clat$mxf == 0)

# Summary of the results
nrow(mxf_res_clat) #number of genes tested
summary(mxf_clat)# Summary of results; upregulated in males with females as the reference
table(mxf_res_clat$mxf)

#Creating a dataframe form table
clat_mxf_table = as.data.frame(table(mxf_res_clat$mxf)) %>%
  rename(Bias = Var1, genes_counts = Freq) %>%
  mutate(Bias = gsub("-1", "Female-biased", Bias), 
         Bias = gsub("1", "Male-biased", Bias), 
         Bias = gsub("0", "Sex-Neutral", Bias)) %>%
  mutate(Species = c(rep("Clat_SB", 3))) %>%
  mutate(Proportion = genes_counts / sum(genes_counts))

# Save results to a CSV file
# write.csv(mxf_res_clat, file = "../3-Output/DEGAnalysis/mxf_genes_clatspecificgenes.csv")
```


####iii EFFECTS OF TISSUE & SEX DIFFERENCES
```{r}
# Perform contrast analysis 
tsxsex_clat = results(dds_clat, contrast = list("tissueS.sexM"), alpha = 0.05) #Baseline Gonad

#Combine results and add a column indicating significance and direction of change
tsxsex_res_clat <- cbind(
  as.data.frame(tsxsex_clat), 
  as.data.frame(tsxsex_clat) %>% 
    mutate(tsxsex_clat = ifelse(padj > 0.05, 0, ifelse(log2FoldChange > 0, 1, -1))) %>% 
    dplyr::select(tsxsex_clat)
  )

tsxsex_res_clat = na.omit(tsxsex_res_clat)
tsxsex_res1_clat = tsxsex_res_clat %>% filter(tsxsex_res_clat$tsxsex != 0 )
tsxsex_res2_clat = tsxsex_res_clat %>% filter(tsxsex_res_clat$tsxsex == 0 )


#Summary of results
summary(tsxsex_clat)
nrow(tsxsex_res_clat)
table(tsxsex_res_clat$tsxsex)
```

##C. Combining data for Cre and Clat
```{r}
combined_speciesgenes = rbind(cre_mxf_table, cre_gxs_table, clat_mxf_table, clat_gxs_table)

# Save results to a CSV file
# write.table(combined_speciesgenes, file = "../3-Output/DEGAnalysis/combined_speciesspecificgenes.txt", row.names = FALSE, col.names = FALSE, quote = FALSE)

```



###############################################
##SUPPLEMENTARY FILES/TABLES##########################
###############################################
#1. Species-biased orthologous genes 
```{r}
crexclat_res #species-biased info
gxs_res #tissue-biased genes
mxf_res #sex-biased 
```
